work on assembler

+++ include/chiplet/chip8compiler.hpp
@@ -46,7 +46,7 @@ public:
     std::string rawErrorMessage() const;
     int errorLine() const;
     int errorCol() const;
-    uint16_t codeSize() const;
+    uint32_t codeSize() const;
     const uint8_t* code() const;
     const std::string& sha1Hex() const;
     std::pair<uint32_t, uint32_t> addrForLine(uint32_t line) const;
